Renamed handlers to actions to follow new conventions
