+++ config/routes.php
@@ -6,6 +6,6 @@ use Zend\Expressive\Application;
 use Zend\Expressive\MiddlewareFactory;
 
 return function (Application $app, MiddlewareFactory $factory, ContainerInterface $container): void {
-    $app->get('/book/{id}/check-out', \App\Action\CheckOutAction::class, 'check-out');
-    $app->get('/book/{id}/check-in', \App\Action\CheckInAction::class, 'check-in');
+    $app->get('/book/{id}/check-out', \App\Handler\CheckOutHandler::class, 'check-out');
+    $app->get('/book/{id}/check-in', \App\Handler\CheckInHandler::class, 'check-in');
 };
