+++ src/App/Handler/CheckInHandlerFactory.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types = 1);
 
-namespace App\Action;
+namespace App\Handler;
 
 use App\Service\Book\FindBookByUuidInterface;
 use Doctrine\ORM\EntityManagerInterface;
@@ -10,17 +10,17 @@ use Interop\Container\ContainerInterface;
 /**
  * @codeCoverageIgnore
  */
-final class CheckOutActionFactory
+final class CheckInHandlerFactory
 {
     /**
      * @param ContainerInterface $container
-     * @return CheckOutAction
+     * @return CheckInHandler
      * @throws \Psr\Container\ContainerExceptionInterface
      * @throws \Psr\Container\NotFoundExceptionInterface
      */
-    public function __invoke(ContainerInterface $container) : CheckOutAction
+    public function __invoke(ContainerInterface $container) : CheckInHandler
     {
-        return new CheckOutAction(
+        return new CheckInHandler(
             $container->get(FindBookByUuidInterface::class),
             $container->get(EntityManagerInterface::class)
         );
