+++ src/App/Handler/CheckOutHandlerFactory.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types = 1);
 
-namespace App\Action;
+namespace App\Handler;
 
 use App\Service\Book\FindBookByUuidInterface;
 use Doctrine\ORM\EntityManagerInterface;
@@ -10,17 +10,17 @@ use Interop\Container\ContainerInterface;
 /**
  * @codeCoverageIgnore
  */
-final class CheckInActionFactory
+final class CheckOutHandlerFactory
 {
     /**
      * @param ContainerInterface $container
-     * @return CheckInAction
+     * @return CheckOutHandler
      * @throws \Psr\Container\ContainerExceptionInterface
      * @throws \Psr\Container\NotFoundExceptionInterface
      */
-    public function __invoke(ContainerInterface $container) : CheckInAction
+    public function __invoke(ContainerInterface $container) : CheckOutHandler
     {
-        return new CheckInAction(
+        return new CheckOutHandler(
             $container->get(FindBookByUuidInterface::class),
             $container->get(EntityManagerInterface::class)
         );
