+++ test/AppTest/Handler/CheckOutHandlerTest.php
@@ -1,9 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace AppTest\Action;
+namespace AppTest\Handler;
 
-use App\Action\CheckInAction;
+use App\Handler\CheckOutHandler;
 use App\Entity\Book;
 use App\Service\Book\Exception\BookNotFound;
 use App\Service\Book\FindBookByUuidInterface;
@@ -17,11 +17,11 @@ use Zend\Diactoros\Response;
 use Zend\Diactoros\ServerRequest;
 
 /**
- * @covers \App\Action\CheckInAction
+ * @covers \App\Handler\CheckOutHandler
  */
-final class CheckInActionTest extends TestCase
+final class CheckOutHandlerTest extends TestCase
 {
-    public function testResponseIs404WhenBookAlreadyStockedThrown() : void
+    public function testResponseIs404WhenBookNotFoundThrown() : void
     {
         $uuid = Uuid::uuid4();
 
@@ -35,10 +35,10 @@ final class CheckInActionTest extends TestCase
         $entityManager = $this->createMock(EntityManagerInterface::class);
         $entityManager->expects(self::never())->method('transactional');
 
-        $action = new CheckInAction($findBookByUuid, $entityManager);
+        $handler = new CheckOutHandler($findBookByUuid, $entityManager);
 
         /** @var Response\JsonResponse $response */
-        $response = $action->process(
+        $response = $handler->process(
             (new ServerRequest(['/']))
                 ->withAttribute('id', $uuid)
                 ->withAttribute(SessionMiddleware::SESSION_ATTRIBUTE, $this->createMock(SessionInterface::class)),
@@ -49,9 +49,10 @@ final class CheckInActionTest extends TestCase
         self::assertSame(404, $response->getStatusCode());
     }
 
-    public function testResponseIs423WhenBookAlreadyCheckedIn() : void
+    public function testResponseIs423WhenBookAlreadyCheckedOut() : void
     {
         $book = Book::fromName('foo');
+        $book->checkOut();
 
         $findBookByUuid = $this->createMock(FindBookByUuidInterface::class);
         $findBookByUuid->expects(self::once())->method('__invoke')->with($book->getId())->willReturn($book);
@@ -59,10 +60,10 @@ final class CheckInActionTest extends TestCase
         $entityManager = $this->createMock(EntityManagerInterface::class);
         $entityManager->expects(self::once())->method('transactional')->willReturnCallback('call_user_func');
 
-        $action = new CheckInAction($findBookByUuid, $entityManager);
+        $handler = new CheckOutHandler($findBookByUuid, $entityManager);
 
         /** @var Response\JsonResponse $response */
-        $response = $action->process(
+        $response = $handler->process(
             (new ServerRequest(['/']))
                 ->withAttribute('id', $book->getId())
                 ->withAttribute(SessionMiddleware::SESSION_ATTRIBUTE, $this->createMock(SessionInterface::class)),
@@ -76,7 +77,6 @@ final class CheckInActionTest extends TestCase
     public function testResponseIs200WhenBookSuccessfullyCheckedOut() : void
     {
         $book = Book::fromName('foo');
-        $book->checkOut();
 
         $findBookByUuid = $this->createMock(FindBookByUuidInterface::class);
         $findBookByUuid->expects(self::once())->method('__invoke')->with($book->getId())->willReturn($book);
@@ -84,10 +84,10 @@ final class CheckInActionTest extends TestCase
         $entityManager = $this->createMock(EntityManagerInterface::class);
         $entityManager->expects(self::once())->method('transactional')->willReturnCallback('call_user_func');
 
-        $action = new CheckInAction($findBookByUuid, $entityManager);
+        $handler = new CheckOutHandler($findBookByUuid, $entityManager);
 
         /** @var Response\JsonResponse $response */
-        $response = $action->process(
+        $response = $handler->process(
             (new ServerRequest(['/']))
                 ->withAttribute('id', $book->getId())
                 ->withAttribute(SessionMiddleware::SESSION_ATTRIBUTE, $this->createMock(SessionInterface::class)),
